Created EnrichedLabel class, which can resize to fit strings of variable length.
Right now, the sizeThatFitsInWidth method goes character by character, so if a graffiti message is long enough to require multiple lines, the string will end mid-word.  I'll update the resizing mechanisms in a later commit to account for space delimiters so we can line break on words instead of characters.