Used HTTPConnection to load nearby graffiti.
This will use the HTTPConnection class to send http requests and use the response to populate the _items array for the tableview.  Wifi on my device isn't working, so I haven't been able to test this.  Possible issues that haven't been thoroughly tested because of wifi problems: 1) HTTPConnection internal working, 2) server implementation/responses, 3) updateFromDictionary methods.  Need to test using wifi, but once we get the HTTPConnectionListener paradigm working, it'll cut out a lot of duplicated code.